Fixed some fields whose sizes were not in line with those in dat.h

+++ project.h
@@ -9,9 +9,9 @@
 typedef struct
 {
 animation_t* animation;
-uint16_t flags;
+uint32_t flags;
 uint16_t friction;
-uint8_t spacing;
+uint16_t spacing;
 uint8_t z_value;
 }car_settings_t;
